pyproject.toml
[project]
name = "kimi-advisor"
version = "0.1.0"
requires-python = ">=3.11"

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
test = [
    "pytest>=8.0",
    "openai>=1.40.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]
